fix: Resolve issue with func_alr not working
This was due to us trying to apply dataframe logic to a numpy array. Thus refactored code to perform specific numpy operations, including removing the chosen column from the numerator matrix so that it can be the denominator vector.

Also, have performed specific operation to subtract the same vector from a matrix.

Lastly, have added logic to get a list of column names to turn the output array back into a dataframe.