docs: Update README to reflect materials
This is to make it more accurate in decribing that this project is concerned with structural break detection in contrast to previously when it focused on time-series analysis and modelling.

Also, deletes the files for the Jupyter book as we don't want to version-control these. Instead, we will create them at the end before we erge with master.